requirements.txt
discord.py==2.3.2
quart==0.19.4
aiosqlite==0.19.0
bcrypt==4.1.2
python-dotenv==1.0.1
hypercorn==0.16.0
itsdangerous==2.1.2
